chore(backend): adapt to new database url

+++ apps/backend/src/models.rs
@@ -871,7 +871,7 @@ pub mod media {
     #[derive(Debug, Serialize, Deserialize, InputObject, Clone)]
     pub struct ProgressUpdateInput {
         pub metadata_id: i32,
-        pub progress: Option<i32>,
+        pub progress: Option<Decimal>,
         pub date: Option<NaiveDate>,
         pub show_season_number: Option<i32>,
         pub show_episode_number: Option<i32>,
@@ -1037,7 +1037,7 @@ pub mod media {
     #[schematic(rename_all = "snake_case")]
     pub struct ImportOrExportMediaItemSeen {
         /// The progress of media done. If none, it is considered as done.
-        pub progress: Option<i32>,
+        pub progress: Option<Decimal>,
         /// The timestamp when started watching.
         pub started_on: Option<DateTimeUtc>,
         /// The timestamp when finished watching.
